pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yt-automation"
version = "0.1.0"
description = "Automated YouTube content pipeline for Roman Urdu channels"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.15.0",
    "pydantic-settings>=2.7.0",
    "gtts>=2.5.0",
    "elevenlabs>=1.0.0",
    "httpx>=0.28.0",
    "google-api-python-client>=2.160.0",
    "google-auth-oauthlib>=1.2.0",
    "Pillow>=11.0.0",
    "rich>=13.9.0",
]

[project.scripts]
yt-auto = "src.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
